chore: moved to dedicated server folder
