+++ src/server/bootstrap/logger-global.ts
@@ -1,4 +1,4 @@
-import createLogger from "~/log";
+import createLogger from "~/server/log";
 
 const logger = createLogger();
