+++ src/server/util/scheduler.ts
@@ -1,0 +1,196 @@
+import type { ScheduledTask } from "node-cron";
+import { schedule as scheduleTask } from "node-cron";
+import moment from "moment-timezone";
+import type {
+  ISchedule,
+  IScheduleConfiguration,
+} from "~/server/database/models/schedule";
+import { getAllEmails as getAllEmailsFromDb } from "~/server/routes/refreshToken";
+import {
+  getAll as getAllSchedulesFromDb,
+  upsert as upsertScheduleInDb,
+  deleteById as deleteScheduleFromDb,
+} from "~/server/routes/schedule";
+import { sendEmail } from "./mailing";
+import { Fleet } from "~/server/util/fleet";
+
+export class Scheduler {
+  private static instance: Scheduler;
+
+  private enabledScheduledTasks: Map<string, ScheduledTask> = new Map();
+  private validEmails: string[] = [];
+
+  private constructor() {}
+
+  public static getInstance(): Scheduler {
+    if (!Scheduler.instance) {
+      Scheduler.instance = new Scheduler();
+    }
+    return Scheduler.instance;
+  }
+
+  private isValidSchedule(schedule: ISchedule): boolean {
+    if (!this.validEmails.includes(schedule.email)) {
+      logger.warn(
+        `Schedule for email ${schedule.email} has no corresponding refresh token.`,
+      );
+      return false; // For now, just skip this schedule. Will re-think later.
+    }
+    if (!schedule.enabled) {
+      logger.info(`Schedule with ID ${schedule.id} is disabled.`);
+      return false;
+    }
+    if (schedule.expires_at && new Date(schedule.expires_at) < new Date()) {
+      logger.warn(`Schedule with ID ${schedule.id} has expired.`);
+      return false;
+    }
+    if (!schedule.configuration || schedule.configuration.length === 0) {
+      logger.warn(
+        `Schedule with ID ${schedule.id} has no configuration defined.`,
+      );
+      return false; // For now, just skip this schedule. Will re-think later if we should allow empty configs at all.
+    }
+    return true;
+  }
+
+  private isValidConfigurationItem(
+    config: IScheduleConfiguration,
+    schedule: ISchedule,
+  ): boolean {
+    if (!config.action || !config.value) {
+      logger.warn(
+        `Invalid configuration for schedule ID ${schedule.id}: ${JSON.stringify(config)}`,
+      );
+      return false;
+    }
+    if (!Fleet.getInstance(schedule.email).getActionMap()[config.action]) {
+      logger.warn(
+        `There is no action defined for ${config.action} in schedule ID ${schedule.id}. Skipping...`,
+      );
+      return false;
+    }
+    return true;
+  }
+
+  async initializeOneSchedule(schedule: ISchedule) {
+    if (!this.isValidSchedule(schedule)) {
+      return;
+    }
+    const task = scheduleTask(
+      schedule.cron,
+      async () => {
+        const now = moment().tz(schedule.timezone);
+        if (now.isAfter(moment(schedule.expires_at))) {
+          logger.warn(`Schedule with ID ${schedule.id} has expired.`);
+          this.enabledScheduledTasks.delete(schedule.id || "");
+          task.destroy();
+          return;
+        }
+        logger.info(
+          `Executing scheduled task for email ${schedule.email} with ID ${schedule.id}`,
+        );
+        try {
+          const products = await Fleet.getInstance(schedule.email)
+            .getEnergyProducts()
+            .then((allProducts) => {
+              return schedule.device_id === "ALL"
+                ? allProducts
+                : allProducts.filter(
+                    (product) => product.id === schedule.device_id,
+                  );
+            });
+
+          for (const product of products) {
+            for (const config of schedule.configuration || []) {
+              if (!this.isValidConfigurationItem(config, schedule)) {
+                continue;
+              }
+              /* eslint-disable no-unexpected-multiline */
+              await Fleet.getInstance(schedule.email)
+                .getActionMap()
+                [config.action](product, parseInt(config.value));
+              /* eslint-enable no-unexpected-multiline */
+            }
+          }
+          logger.info(
+            `Executed scheduled task for email ${schedule.email} with ID ${schedule.id} successfully`,
+          );
+          upsertScheduleInDb({
+            id: schedule.id,
+            lastRunTime: now.toDate(),
+            nextRunTime: task.getNextRun() || undefined,
+            lastSuccessTime: now.toDate(),
+          });
+        } catch (error: any) {
+          const lastError = error.message || "Unknown error";
+          logger.error(
+            `Error executing scheduled task for email ${schedule.email} with ID ${schedule.id}: ${lastError}`,
+          );
+          sendEmail(
+            "Powerwall Notification",
+            `[${new Date().toLocaleString()}] ${lastError}`,
+            schedule.email,
+          );
+          upsertScheduleInDb({
+            id: schedule.id,
+            lastRunTime: now.toDate(),
+            nextRunTime: task.getNextRun() || undefined,
+            lastError,
+            lastErrorTime: now.toDate(),
+          });
+        }
+      },
+      { timezone: schedule.timezone },
+    );
+    this.enabledScheduledTasks.set(schedule.id || "", task);
+  }
+
+  async initialize() {
+    this.validEmails = await getAllEmailsFromDb();
+    this.enabledScheduledTasks.clear();
+    for (const schedule of await getAllSchedulesFromDb()) {
+      await this.initializeOneSchedule(schedule);
+    }
+  }
+
+  async stopAll() {
+    for (const [id, task] of this.enabledScheduledTasks.entries()) {
+      logger.info(`Stopping scheduled task with ID ${id}`);
+      task.stop();
+      this.enabledScheduledTasks.delete(id);
+    }
+    logger.info("All scheduled tasks stopped.");
+  }
+
+  async startAll() {
+    for (const [id, task] of this.enabledScheduledTasks.entries()) {
+      logger.info(`Starting scheduled task with ID ${id}`);
+      task.start();
+    }
+    logger.info("All scheduled tasks started.");
+  }
+
+  async upsert(schedule: ISchedule) {
+    if (!this.isValidSchedule(schedule)) {
+      return;
+    }
+    const existingTask = this.enabledScheduledTasks.get(schedule.id || "");
+    if (existingTask) {
+      logger.info(`Updating existing scheduled task with ID ${schedule.id}`);
+      existingTask.stop();
+      this.enabledScheduledTasks.delete(schedule.id || "");
+    }
+    await this.initializeOneSchedule(schedule);
+    return upsertScheduleInDb(schedule);
+  }
+
+  async delete(scheduleId: string) {
+    const task = this.enabledScheduledTasks.get(scheduleId);
+    if (task) {
+      logger.info(`Deleting scheduled task with ID ${scheduleId}`);
+      task.stop();
+      this.enabledScheduledTasks.delete(scheduleId);
+    }
+    return deleteScheduleFromDb(scheduleId);
+  }
+}
